reorganize tests and make them inline namespace
